+++ src/Exception/ServerMessageThrower.php
@@ -109,6 +109,7 @@ class ServerMessageThrower
         'Checkpoint\AcknowledgeFormException'                => ['AcknowledgeForm'],
         'Checkpoint\VerifyEmailCodeFormException'            => ['VerifyEmailCodeForm'],
         'Checkpoint\VerifySMSCodeFormException'              => ['VerifySMSCodeForm'],
+        'Checkpoint\LegacyForceSetNewPasswordFormException'  => ['LegacyForceSetNewPasswordForm'],
     ];
 
     /**